- develop : manage dependency injection with network, repo and use case modules, create dashboard activity and view model

+++ domain/src/main/java/com/forecast/domain/entity/Coord.kt
@@ -1,6 +1,6 @@
 package com.forecast.domain.entity
 
 data class Coord(
-    val lat: Int,
+    val lat: Double,
     val lon: Double
 )
